# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gin+gorm 后台API",
        "title": "gin+gorm 测试swagger",
        "contact": {
            "name": "go-swagger文档",
            "url": "https://github.com/swaggo/swag/blob/master/README_zh-CN.md"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": "v1.0"
    },
    "host": "222.196.37.70:8080",
    "paths": {
        "/api/addFriendAppli": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "好友/群聊申请模块"
                ],
                "summary": "新增好友申请",
                "parameters": [
                    {
                        "description": "申请表单",
                        "name": "ApplicationF",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.ApplicationF"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/addGroupAppli": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "好友/群聊申请模块"
                ],
                "summary": "新增群聊申请",
                "parameters": [
                    {
                        "description": "申请表单",
                        "name": "ApplicationF",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.ApplicationF"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/dealFriendAppli": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "好友/群聊申请模块"
                ],
                "summary": "好友处理申请",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "申请Id",
                        "name": "aid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "是否同意",
                        "name": "flag",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/dealGroupAppli": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "好友/群聊申请模块"
                ],
                "summary": "群主处理申请",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "申请Id",
                        "name": "aid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "是否同意",
                        "name": "flag",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/deleteFriends": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "好友管理模块"
                ],
                "summary": "删除好友",
                "parameters": [
                    {
                        "type": "string",
                        "description": "好友id",
                        "name": "fid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/getAppliMsg": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "好友/群聊申请模块"
                ],
                "summary": "申请人查询申请信息",
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/getfriends": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "好友管理模块"
                ],
                "summary": "获取好友列表",
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/userGetAppliMsg": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "好友/群聊申请模块"
                ],
                "summary": "好友接受者查询申请信息",
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "service.ApplicationF": {
            "type": "object",
            "required": [
                "applicantid",
                "applicantname",
                "reviewerid",
                "reviewername"
            ],
            "properties": {
                "applicantid": {
                    "type": "integer"
                },
                "applicantname": {
                    "type": "string"
                },
                "reviewerid": {
                    "type": "integer"
                },
                "reviewername": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}